mix: tidy up docs, remove Blade directive, use Blade view
